+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-# 2018-11-08 0.4.0
+# 0.5.0 (2019-01-02)
+
+* [BREAKING] Move `health_check_server` to `health_check.enabled` and `health_check.address`
+* Add `tag_metric` to targets
+* Add `latency_ms` to probe results
+* Add Prometheus support, configure using `metrics.enabled`, `metrics.address`
+* Removed `hyper` as a direct dependency
+
+# 0.4.0 (2018-11-08)
 
 * [BREAKING] Moved `config.alert` to `config.alert.email`
 * Add notifications to dashboard: `http://rcanary.example.com?notifications=true`